video: Silence warning from compiler.
`SyncTargetBin` and `i` are both `guint`s, which is an unsigned integer
type (Glib-flavoured).  Not sure how big a `guint` can get, but if
`i` ever becomes bigger than `SyncTargetBin`, you're going to get it
wrapping around that maximum value.

So cast it to a signed type.  `gint` seems the logical choice here.