pcm: Add channel selection
Some radio interfaces use a stereo input to support two mono receivers
(e.g. NWDR UDRC-II and DRAWS both have this).  This allows for selection
of one channel over the other, or the option of mixing both channels
together.

The default is to use the left channel, as that matches what the audio
subsystem did prior to this feature being added.